+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Read `release_notes.md` for commit level details.
 
 ## [Unreleased]
 ### Enhancements
+
+### Bug fixes
+
+### Deprecations
+
+## [2.3.0] - 2019-01-07
+### Enhancements
 - Use `Base64.strict_encode64` when this client sends `Base64` encoded data to server
     - Follows RFC 4648 format. It should not affect server side which is front Appium node server
     - Continues to decode base 64 data following `decode64` to accept RFC 2045 format